Avoid caching of the get-rev-count derivation
The derivation is used in install-image.nix to determine the
Git revision of a checked-out nixpkgs repo.  It will be realized
only once and all subsequent runs will use the cached version even
if the checked-out revision has changed.  Caching is avoided by
adding a dummy attribute that contains the current time to force
re-evaluation for every invocation.