Create a default NixOS configuration if nixosConfigDir is null
Creating the default with nixos-generate-config during installation
has the disadvantage that the packages required by it are not part of
the system closure when the image is created.  It also introduces a
kind non-deterministic behaviour.  When nixosConfigDir is left empty
(null), a minimalistic configuration.nix is created to select
gummiboot as boot loader and set the initial root password to make the
system accessible via the console.