Major overhaul
The system is now based on NixOS 18.03 and uses some features that
were not available on 16.03.

The install image is no longer built in a VM, which greatly speeds up
the process.  Transfer of the system closure and channel is done more
naturally with `nixos-install`.

The installer has been simplified

   * Interface auto-detection by GRUB and the Linux kernel made the
     efinetDHCPInterface and linuxPnPInterface options obsolete

   * grub.cfg is now loaded vie TFTP rather than built into the
     boot loader

   * Support of client-specific GRUB hooks

   * Replacement of plain chroot with nixos-enter has simplified
     the installation script

   * Removal of interface configuration (useDHCP,
     staticInterfaceFromDHCP options), default config enables
     DHCP for all interfaces