Fix sandbox builds
Starting with NixOS 18.09, the default of nix.useSandbox was changed
to true, which broke the logic around the readLink derivation.  The
system now expects the installImage.nixpkgs.path option to be an
object in the Nix store to solve this problem.

The URL of the binary cache of the constructed channel is now always
taken from the installImage.binaryCacheURL option.